+++ ajek-framework/src/x-stdlib.cpp
@@ -749,7 +749,7 @@ void xStrCopy(char* dest, size_t destLen, const char* src, size_t srcLen)
 // x-chrono.h things  (not in a mood to put these in their own module, yet...)
 
 host_tick_t		HostClockTick::s_ticks_per_second;
-float			HostClockTick::s_ticks_per_second_f;
+double			HostClockTick::s_ticks_per_second_f;
 
 void HostClockTick::Init()
 {
